convert_to_double_add_inputs

+++ kmean@c/vector.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-#include "vector.h" // Include the Vector header file
+#include "vector.h" 
 
 
 // Function to create a new Vector
-Vector createVector(int dimension, float *values) {
+Vector createVector(int dimension, double *values) {
     Vector vec;
     vec.dimension = dimension;
     vec.centroid = -1; // Initialize centroid to -1 (indicating not assigned)
     
     if (values == NULL)
-        vec.components = (float *)calloc(dimension, sizeof(float)); // Allocate memory for components
+        vec.components = (double *)calloc(dimension, sizeof(double)); // Allocate memory for components
     else {
-        vec.components = (float *)malloc(dimension * sizeof(float)); // Allocate memory for components
+        vec.components = (double *)malloc(dimension * sizeof(double)); // Allocate memory for components
         if (vec.components == NULL) {
             fprintf(stderr, "Memory allocation failed\n");
             exit(EXIT_FAILURE);
@@ -42,7 +42,7 @@ Vector add(Vector vec1, Vector vec2) {
 }
 
 // Function to perform scalar multiplication of a vector
-Vector multiplyScalar(Vector vec, float scalar) {
+Vector multiplyScalar(Vector vec, double scalar) {
     Vector result = createVector(vec.dimension, NULL); // Create a result vector
     // Perform scalar multiplication component-wise
     for (int i = 0; i < vec.dimension; i++) {
@@ -52,16 +52,16 @@ Vector multiplyScalar(Vector vec, float scalar) {
 }
 
 // Function to calculate the Euclidean distance between two vectors
-float euclidean_distance(Vector vec1, Vector vec2) {
+double euclidean_distance(Vector vec1, Vector vec2) {
     if (vec1.dimension != vec2.dimension) {
         fprintf(stderr, "Vectors must have the same dimension for calculating Euclidean distance\n");
         exit(EXIT_FAILURE);
     }
     
-    float sum = 0.0;
+    double sum = 0.0;
     // Calculate the sum of squared differences component-wise
     for (int i = 0; i < vec1.dimension; i++) {
-        float diff = vec1.components[i] - vec2.components[i];
+        double diff = vec1.components[i] - vec2.components[i];
         sum += diff * diff;
     }
     return sqrt(sum); // Return the square root of the sum
@@ -79,4 +79,3 @@ void printVector(Vector vec) {
     }
     printf(")\n");
 }
-
